Messy and broken HTTP proxy code.
When trying to access a blocked website that is now allowed, our filter will act as an HTTP proxy to give the user access. This code is setup to proxy yahoo.com. The proxy has difficulty determining when the client or server closes the connection (possibly due to it not having proper support for persistent connections).